kehamilan role

+++ application/views/data/ibu_hamil/kehamilan/index.php
@@ -8,14 +8,16 @@
             <div class="col-md-10 col-xs-12 align-self-center">
                <h6 class="m-0 font-weight-bold text-primary mb-2"><?= $title ?></h6>
             </div>
-            <div class="col-md-2 col-xs-12">
-               <a type="button" class="btn btn-success float-right ml-2 btn-block" href="<?= base_url('data/ibu_hamil/kehamilan/pdf') ?>">
-                  <i class="fas fa-print"></i> PDF
-               </a>
-               <button type="button" class="btn btn-primary float-right btn-block" data-toggle="modal" data-target="#addModal">
-                  <i class="fas fa-plus"></i> Tambah
-               </button>
-            </div>
+            <?php if ($role == 7) : ?>
+               <div class="col-md-2 col-xs-12">
+                  <a type="button" class="btn btn-success float-right ml-2 btn-block" href="<?= base_url('data/ibu_hamil/kehamilan/pdf') ?>">
+                     <i class="fas fa-print"></i> PDF
+                  </a>
+                  <button type="button" class="btn btn-primary float-right btn-block" data-toggle="modal" data-target="#addModal">
+                     <i class="fas fa-plus"></i> Tambah
+                  </button>
+               </div>
+            <?php endif; ?>
          </div>
       </div>
       <div class="card-body">
@@ -36,7 +38,9 @@
                      <th>Jumlah Lahir Mati</th>
                      <th>Jarak Persalinan Terakhir</th>
                      <th>Jenis Persalinan Terakhir</th>
-                     <th>Aksi</th>
+                     <?php if ($role == 7) : ?>
+                        <th>Aksi</th>
+                     <?php endif; ?>
                   </tr>
                </thead>
                <tbody>
@@ -53,10 +57,12 @@
                         <td><?= $field['jml_lahir_mati'] ?></td>
                         <td><?= $field['jarak_persalinan_terakhir'] ?></td>
                         <td><?= $field['jenis_persalinan_terakhir'] ?></td>
-                        <td>
-                           <button type="button" class="btn btn-block btn-primary" data-toggle="modal" data-target="#editModal" onclick="getData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Edit</button>
-                           <button type="button" class="btn btn-block btn-danger" data-toggle="modal" data-target="#deleteModal" onclick="deleteData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Hapus</button>
-                        </td>
+                        <?php if ($role == 7) : ?>
+                           <td>
+                              <button type="button" class="btn btn-block btn-primary" data-toggle="modal" data-target="#editModal" onclick="getData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Edit</button>
+                              <button type="button" class="btn btn-block btn-danger" data-toggle="modal" data-target="#deleteModal" onclick="deleteData(<?= htmlspecialchars(json_encode($field), ENT_QUOTES, 'UTF-8') ?>)">Hapus</button>
+                           </td>
+                        <?php endif; ?>
                      </tr>
                   <?php endforeach; ?>
                </tbody>
